backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
snowflake-connector-python>=3.5.0
python-dotenv>=1.0.0
cryptography>=41.0.0
